remove qt dependency from v2 again
